+++ artifex/models/base_model.py
@@ -27,16 +27,13 @@ class BaseModel(ABC):
     """
     Base class for all models in the artifex package.
     """
+    
+    def __init__(self, synthex: Synthex):
+        self._synthex_val: Synthex = synthex
+        self._tokenizer_val: PreTrainedTokenizerBase = PreTrainedTokenizerBase()
+        self._model_val: Optional[PreTrainedModel] = None
 
     ##### Abstract properties #####
-    
-    @property
-    @abstractmethod
-    def _synthex(self) -> Synthex:
-        """
-        An instance of the Synthex class.
-        """
-        pass
 
     @property
     @abstractmethod
@@ -48,28 +45,18 @@ class BaseModel(ABC):
     
     @property
     @abstractmethod
-    def _model(self) -> Optional[PreTrainedModel]:
-        """
-        A trainable model to perform inference with. It may be None if the task characteristics make
-        it impossible to select a model at instantiation time (e.g., in a multiclass classification task,
-        where the number of classes is not known upfront, selecting a model at instantiation is impossible).
-        """
-        pass
-    
-    @property
-    @abstractmethod
-    def _tokenizer(self) -> PreTrainedTokenizerBase:
-        """
-        The tokenizer used during training and inference.
-        """
-        pass
-    
-    @property
-    @abstractmethod
     def _token_keys(self) -> list[str]:
         """
         The keys in the dataset that contain the text to be tokenized. These are used to tokenize
         the dataset during training and inference.
+        """
+        pass
+    
+    @property
+    @abstractmethod
+    def _base_model_name(self) -> str:
+        """
+        The name of the base model to use for the model.
         """
         pass
     
@@ -176,6 +163,38 @@ class BaseModel(ABC):
         pass
     
     ##### Methods #####
+    
+    @property
+    def _synthex(self) -> Synthex:
+        """
+        The Synthex instance used to generate synthetic data for training the model.
+        """
+        return self._synthex_val
+    
+    @property
+    def _tokenizer(self) -> PreTrainedTokenizerBase:
+        """
+        The tokenizer used to preprocess text data for the model.
+        """
+        return self._tokenizer_val
+    
+    @property
+    def _model(self) -> Optional[PreTrainedModel]:
+        """
+        A trainable model to perform inference with. It may be None if the task characteristics 
+        make it impossible to select a model at instantiation time (e.g., in a multiclass 
+        classification task, where the number of classes is not known upfront).
+        """
+        return self._model_val
+    
+    @_model.setter
+    def _model(self, model: PreTrainedModel) -> None:
+        """
+        Set the trainable model to perform inference with.
+        Args:
+            model (PreTrainedModel): The model to set.
+        """
+        self._model_val = model
     
     @staticmethod
     def _sanitize_output_path(output_path: Optional[str] = None) -> str: