+++ artifex/models/intent_classifier/__init__.py
@@ -1,14 +1,11 @@
 from synthex import Synthex
-from synthex.models import JobOutputSchemaDefinition
 from transformers import AutoTokenizer, AutoModelForSequenceClassification, PreTrainedTokenizerBase, \
     PreTrainedModel
 from datasets import ClassLabel # type: ignore
 
-from .models import IntentClassifierInstructions
-
 from artifex.models.nclass_classification_model import NClassClassificationModel
-from artifex.config import config
 from artifex.core import auto_validate_methods
+from artifex.config import config
 
 
 @auto_validate_methods
@@ -25,12 +22,8 @@ class IntentClassifier(NClassClassificationModel):
             synthex (Synthex): An instance of the Synthex class to generate the synthetic data used to train the model.
         """
         
-        super().__init__()
-        self._synthex_val: Synthex = synthex
-        self._synthetic_data_schema_val: JobOutputSchemaDefinition = {
-            "text": {"type": "string"},
-            "labels": {"type": "string"},
-        }
+        super().__init__(synthex)
+        self._base_model_name_val: str = config.INTENT_CLASSIFIER_HF_BASE_MODEL
         self._system_data_gen_instr: list[str] = [
             "The 'text' field should contain text that has a specific intent or objective.",
             "The 'labels' field should contain a label indicating the intent or objective of the 'text'.",
@@ -38,49 +31,18 @@ class IntentClassifier(NClassClassificationModel):
             "This is a list of the allowed 'labels' and 'text' pairs: "
         ]
         self._model_val: PreTrainedModel = AutoModelForSequenceClassification.from_pretrained( # type: ignore
-            config.INTENT_CLASSIFIER_HF_BASE_MODEL
+            self._base_model_name
         )
         self._tokenizer_val: PreTrainedTokenizerBase = AutoTokenizer.from_pretrained( # type: ignore
-            config.INTENT_CLASSIFIER_HF_BASE_MODEL
+            self._base_model_name
         )
-        self._token_keys_val: list[str] = ["text"]
         self._labels_val: ClassLabel = ClassLabel(
             names=list(self._model_val.config.id2label.values()) # type: ignore
         )
-
-    @property
-    def _synthex(self) -> Synthex:
-        return self._synthex_val
-
-    @property
-    def _synthetic_data_schema(self) -> JobOutputSchemaDefinition:
-        return self._synthetic_data_schema_val
-    
-    @property
-    def _tokenizer(self) -> PreTrainedTokenizerBase:
-        return self._tokenizer_val
-    
-    @property
-    def _token_keys(self) -> list[str]:
-        return self._token_keys_val
-    
-    def _parse_user_instructions(self, user_instructions: IntentClassifierInstructions) -> list[str]:
-        """
-        Turn the data generation job instructions provided by the user from a IntentClassifierInstructions object 
-        into a list of strings that can be used to generate synthetic data through Synthex.   
-        Args:
-            user_instructions (IntentClassifierInstructions): Instructions provided by the user for generating synthetic data.
-            extra_instructions (list[str]): A list of additional instructions to include in the data generation.
-        Returns:
-            list[str]: A list of complete instructions for generating synthetic data.
-        """
         
-        out: list[str] = []
-        
-        for class_name, description in user_instructions.items():
-            out.append(f"{class_name}: {description}")
-        
-        return out
+    @property
+    def _base_model_name(self) -> str:
+        return self._base_model_name_val
     
     def _get_data_gen_instr(self, user_instr: list[str]) -> list[str]:
         return self._system_data_gen_instr + user_instr