+++ artifex/models/intent_classifier/models.py
@@ -1,4 +1,0 @@
-from artifex.core import ClassificationClassName
-    
-
-IntentClassifierInstructions = dict[ClassificationClassName, str]